% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencorpfunc.R
\name{fingerprint.func}
\alias{fingerprint.func}
\title{Function to clean a given string based on the fingerprint method.}
\usage{
fingerprint.func(x)
}
\arguments{
\item{x}{a vector of string.}
}
\description{
Function to clean a given string based on the fingerprint method.
}
\details{
The process that generates the key from a string value is the following:

\itemize{
\item remove leading and trailing whitespace
\item change all characters to their lowercase representation
\item remove all punctuation and control characters
\item split the string into whitespace-separated tokens
\item sort the tokens and remove duplicates
\item join the tokens back together
\item normalize extended western characters to their ASCII representation
}
}
\note{
If the entry is a numeric, it will be converted to a string character
}

