% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencorpfunc.R
\name{ropencorp}
\alias{ropencorp}
\title{Query opencorporate website}
\usage{
ropencorp(term, nb.page = 2, credential = NULL, country = NULL)
}
\arguments{
\item{term}{Term to query on the open corporate website}

\item{nb.page}{Number of pages to query}

\item{credential}{Credential, for the professionnal version}

\item{country}{List or vector of countries to query}
}
\description{
Query opencorporate website
}

