% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencorpfunc.R
\name{scrape.dt}
\alias{scrape.dt}
\title{Main function to scrape the API}
\usage{
scrape.dt(x)
}
\arguments{
\item{x}{a string to search through the API}
}
\description{
Main function to scrape the API
}
\note{
This is the core funciton of the package. All the rest is a wrapper for that one
}

