% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world.states}
\alias{world.states}
\title{Coordinate of the juridiction code.}
\format{A data table with 299 rows and 4 columns:
\describe{
  \item{jurisdiction.code}{Abreviation of the juridiction code appearing on the open corporate web site}
  \item{lat}{latitude}
  \item{lng}{longitude}
  \item{state.name}{Name of the state}
}}
\source{
handmade compilation from different sources.
}
\usage{
world.states
}
\description{
A dataset containing the coordinate associated to the juridiction code.
}
\keyword{datasets}

