% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencorpfunc.R
\name{fingerprint.func}
\alias{fingerprint.func}
\title{Fingerprint function to clean the character}
\usage{
fingerprint.func(x)
}
\arguments{
\item{x}{a vector of strings.}
}
\description{
Fingerprint function to clean the character
}
\note{
If the entry is a numeric vector, it will be converted to a string character
}

