% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencorpfunc.R
\name{get.comp.number}
\alias{get.comp.number}
\title{Query opencorporate website to search a given company number in a given jurisdiction}
\usage{
get.comp.number(company.number, jurisdiction.code, token = NULL)
}
\arguments{
\item{company.number}{A company number or a vector of comapny numbers.}

\item{jurisdiction.code}{A jurisdiction code or a vector of jurisdiction code which match the length of company.number.}

\item{token}{Token for the owner of an opencorporate account}
}
\value{
The returned object is a list of four data.table if the data.table package is attached. Otherwise, it is a list of four data.frame.
 Here, using the data.table package speed up the script.

\itemize{
\item company.id.dt: Company details
\item officers.comp.dt: List of officers. Link with the previous table withcompany.number and jurisdiction.code
\item identifiers.dt: Identifiers
\item industry.code.dt: Industry code. Link with the previous tables with company.number and jurisdiction.code
}
}
\description{
Query opencorporate website to search a given company number in a given jurisdiction
}
\details{
The returned object is two data.table if the data.table package is attached. Otherwise, it is two data.frame.
 Here, using the data.table package speed up the script.
}
\examples{
   get.comp.number(c("23286336", "30760997"), c("ro", "ro"))

}

