% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencorpfunc.R
\name{get.officers}
\alias{get.officers}
\title{Query opencorporate website}
\usage{
get.officers(term, nb.page = 20, token = NULL, country = NULL,
  ret.score = FALSE)
}
\arguments{
\item{term}{Term to query on the open corporate website}

\item{nb.page}{Number of pages to query}

\item{token}{Token for the owner of an opencorporate account}

\item{country}{List or vector of countries to query}

\item{ret.score}{Return the top results by score. Otherwise, the results are returned by alphabetical order.}
}
\description{
Query opencorporate website
}
\details{
The returned object is two data.table if the data.table package is attached. Otherwise, it is two data.frame.
 Here, using the data.table package speed up the script.
}
\note{
more on the API documentation: https://api.opencorporates.com/documentation/API-Reference#account_status
}
\examples{
  get.officers("Vincent de Rivaz", nb.page = 2)

}

