% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropencorpfunc.R
\name{ropencorp}
\alias{ropencorp}
\title{Query opencorporate website}
\usage{
ropencorp(term, nb.page = 2, token = NULL, country = NULL)
}
\arguments{
\item{term}{Term to query on the open corporate website}

\item{nb.page}{Number of pages to query}

\item{token}{Token for the owner of an opencorporate account}

\item{country}{List or vector of countries to query}
}
\description{
Query opencorporate website
}
\note{
The returned object is a data.table if the data.table package is attached. Otherwise, it is a data.frame. Here, using the data.table package speed up the script.
}
\examples{
  ropencorp("gold holding", nb.page = 2)

}

